rewrite a little
