+++ src/extra_fields/mod.rs
@@ -14,7 +14,7 @@ pub(crate) use zip64_extended_information::Zip64ExtendedInformation;
 
 // re-export
 pub use extended_timestamp::ExtendedTimestamp;
-pub use extra_field::ExtraField;
+pub use extra_field::{ExtraField, ExtraFields};
 pub use ntfs::Ntfs;
 pub use zipinfo_utf8::UnicodeExtraField;
