+++ src/read/mod.rs
@@ -3,7 +3,7 @@
 use crate::compression::CompressionMethod;
 use crate::cp437::FromCp437;
 use crate::datetime::DateTime;
-use crate::extra_fields::ExtraField;
+use crate::extra_fields::{ExtraField, ExtraFields};
 use crate::format::flags::ZipFlags;
 use crate::result::{ZipError, ZipResult, invalid};
 use crate::spec::{CentralDirectoryEndInfo, DataAndPosition, FixedSizeBlock, ZipCentralEntryBlock};
@@ -531,7 +531,8 @@ fn central_header_to_zip_file_inner<R: Read>(
         aes_extra_data_start: 0,
         extra_fields: Vec::new(),
     };
-    parse_extra_field(&mut result, &mut file_name_raw)?;
+    let extra = ExtraFields::parse(&mut extra_field, &mut result, &mut file_name_raw)?;
+    result.extra_fields = extra;
 
     // Account for shifted zip offsets.
     result.header_start = result
@@ -540,119 +541,6 @@ fn central_header_to_zip_file_inner<R: Read>(
         .ok_or(invalid!("Archive header is too large"))?;
 
     Ok((result, file_name_raw))
-}
-
-pub(crate) fn parse_extra_field(
-    file: &mut ZipFileData,
-    file_name_raw: &mut Vec<u8>,
-) -> ZipResult<()> {
-    let mut extra_field = file.extra_field.clone();
-    let mut central_extra_field = file.central_extra_field.clone();
-    for field_group in [&mut extra_field, &mut central_extra_field] {
-        let Some(extra_field) = field_group else {
-            continue;
-        };
-        let mut modified = false;
-        let mut processed_extra_field = vec![];
-        let len = extra_field.len();
-        let mut reader = io::Cursor::new(&**extra_field);
-
-        let mut position = reader.position();
-        while position < len as u64 {
-            let old_position = position;
-            let remove =
-                parse_single_extra_field(file, &mut reader, position, false, file_name_raw)?;
-            position = reader.position();
-            if remove {
-                modified = true;
-            } else {
-                let field_len = (position - old_position) as usize;
-                let write_start = processed_extra_field.len();
-                reader.seek(SeekFrom::Start(old_position))?;
-                processed_extra_field.extend_from_slice(&vec![0u8; field_len]);
-                if let Err(e) = reader
-                    .read_exact(&mut processed_extra_field[write_start..(write_start + field_len)])
-                {
-                    if e.kind() == io::ErrorKind::UnexpectedEof {
-                        return Err(invalid!("Extra field content exceeds declared length"));
-                    }
-                    return Err(e.into());
-                }
-            }
-        }
-        if modified {
-            *field_group = Some(Arc::from(processed_extra_field.into_boxed_slice()));
-        }
-    }
-    file.extra_field = extra_field;
-    file.central_extra_field = central_extra_field;
-    Ok(())
-}
-
-pub(crate) fn parse_single_extra_field<R: Read>(
-    file: &mut ZipFileData,
-    reader: &mut R,
-    bytes_already_read: u64,
-    disallow_zip64: bool,
-    file_name_raw: &mut Vec<u8>,
-) -> ZipResult<bool> {
-    let parsed_extra_field = ExtraField::parse(reader, file)?;
-    let Some(parsed_extra_field) = parsed_extra_field else {
-        return Ok(false);
-    };
-    match parsed_extra_field {
-        // Zip64 extended information extra field
-        ExtraField::Zip64ExtendedInformation {
-            uncompressed_size,
-            compressed_size,
-            header_start,
-        } => {
-            if disallow_zip64 {
-                return Err(invalid!("Can't write a custom field using the ZIP64 ID"));
-            }
-            file.large_file = true;
-            file.uncompressed_size = uncompressed_size;
-            file.compressed_size = compressed_size;
-            file.header_start = header_start;
-            return Ok(true);
-        }
-        ExtraField::Ntfs(ntfs) => {
-            // NTFS extra field
-            file.extra_fields.push(ExtraField::Ntfs(ntfs));
-        }
-        ExtraField::AeXEncryption {
-            aes_mode,
-            aes_vendor_version,
-            compression_method,
-        } => {
-            file.aes_mode = Some((aes_mode, aes_vendor_version));
-            file.compression_method = compression_method;
-            file.aes_extra_data_start = bytes_already_read;
-        }
-        ExtraField::ExtendedTimestamp(extended_timestamp) => {
-            file.extra_fields
-                .push(ExtraField::ExtendedTimestamp(extended_timestamp));
-        }
-        ExtraField::UnicodeComment(unicode_comment) => {
-            // Info-ZIP Unicode Comment Extra Field
-            // APPNOTE 4.6.8 and https://libzip.org/specifications/extrafld.txt
-            file.file_comment = String::from_utf8(
-                unicode_comment
-                    .unwrap_valid(file.file_comment.as_bytes())?
-                    .into_vec(),
-            )?
-            .into();
-        }
-        ExtraField::UnicodePath(unicode_path) => {
-            // Info-ZIP Unicode Path Extra Field
-            // APPNOTE 4.6.9 and https://libzip.org/specifications/extrafld.txt
-            let file_name = unicode_path.unwrap_valid(file_name_raw)?;
-            *file_name_raw = file_name.into_vec();
-            file.flags |= ZipFlags::LanguageEncoding.as_u16();
-        }
-        ExtraField::Unknown(_) => {}
-    }
-    Ok(false)
 }
 
 /// A trait for exposing file metadata inside the zip.
